+++ src/panel_material_ui/pane/base.py
@@ -6,6 +6,7 @@ from typing import Any
 import param
 from panel.links import Callback
 from panel.pane import HoloViews, Markdown
+from panel.viewable import Child
 
 from ..base import COLORS, MaterialComponent, ThemedTransform, TooltipTransform
 from ..widgets import DatetimeInput, DiscreteSlider, EditableFloatSlider, EditableIntSlider, FloatSlider, IntSlider, Select
@@ -123,6 +124,60 @@ class ClickablePaneBase(MaterialPaneBase):
             if val is not None:
                 callbacks[k] = val
         return Callback(self, code=callbacks, args=args)
+
+
+class Badge(MaterialComponent):
+    """
+    The `Badge` generates a small badge to the top-right (by default)
+    of its child element. Badges are commonly used to display
+    notification counts, status indicators, or short labels overlaid
+    on icons, avatars, or buttons.
+
+    :References:
+
+    - https://panel-material-ui.holoviz.org/reference/panes/Badge.html
+    - https://mui.com/material-ui/react-badge/
+
+    :Example:
+
+    >>> Badge(IconButton(icon="mail"), badge_content=4, color="primary")
+    """
+
+    anchor_origin = param.Dict(default=None, doc="""
+        The anchor position of the badge. Accepts a dictionary with
+        keys 'vertical' ('top' or 'bottom') and 'horizontal'
+        ('left' or 'right').""")
+
+    badge_content = param.Parameter(default=0, doc="""
+        The content rendered within the badge. Typically an integer
+        count but can be a short string.""")
+
+    color = param.Selector(default="primary", objects=COLORS, doc="""
+        The color of the badge.""")
+
+    max = param.Integer(default=99, bounds=(0, None), doc="""
+        Maximum count to display. Values above this show as
+        'max+' (e.g. '99+').""")
+
+    object = Child(doc="""
+        The child component to wrap with the badge.""")
+
+    overlap = param.Selector(default="rectangular", objects=["rectangular", "circular"], doc="""
+        Wrapped shape the badge should overlap.""")
+
+    show_zero = param.Boolean(default=False, doc="""
+        Whether to display the badge when badge_content is zero.""")
+
+    variant = param.Selector(default="standard", objects=["dot", "standard"], doc="""
+        The variant of the badge. Use 'dot' for a small dot
+        indicator without content.""")
+
+    _esm_base = "Badge.jsx"
+
+    def __init__(self, object=None, **params):
+        if object is not None:
+            params["object"] = object
+        super().__init__(**params)
 
 
 class Avatar(ClickablePaneBase):
